--- Harlowe/Code Template/Settings/Settings Template.tw
@@ -1,13 +1,0 @@
-:: StoryTitle
-Harlowe Settings
-
-:: StoryData
-{
-  "ifid": "37BFE9B3-93E0-4A2C-B74E-3FAEF445DBC3",
-  "format": "Harlowe",
-  "format-version": "3.3.9",
-  "start": "Untitled Passage",
-  "zoom": 1
-}
-
-
