+++ README.md
@@ -1,9 +1,11 @@
-# Twine-Template
-This Repo includes Twine and Twine related templates and code.\
-Though it (will/does) include more assets for the SugarCube format.
+# ========* Twine Templates *========
+This Repo includes Twine and Twine related templates and code.
 
 The directory of this repository is organised as follow:
-* Harlowe (TBA)
+* Chapbook (TBD)
+* Harlowe
+  * Code Templates
+    * Save System and Settings (also a Guide)
 * SugarCube
   * Code Templates
     * Character Creator
@@ -16,16 +18,31 @@ The directory of this repository is organised as follow:
     * VN-like RPG
 * Ready-to-Use Tweego Folder
 
-## Harlowe Templates (TBA)
+## FORE-WARNING
+**All the templates have the same IFID (per format)!**\
+You will need to change this IFID if you are planning on editing the file for your project.\
+I have a whole explanation on how to do so [here](ChangeIFID.md) or on [itch.io](https://manonamora.itch.io/twine-sugarcube-templates/devlog/568346/help-change-the-ifid-of-a-template).
+
+## Harlowe Templates
+The templates included in this folder were made in Harlowe 3.3.9. They can be open with any version of Twine 2 that includes Harlowe 3.3.9.\
+Using those templates with any older or newer version of Harlowe may break the code.
+
+The templates are organised in 2 categories:
+* Code Templates (saves, settings, etc...)
+* UI Temapltes (interface change) -> TBA
+
+You can find a more comprehensive list of my Harlowe templates, along with their specs [here](Harlowe/DocsHarlowe.md).
 
 ## SugarCube Templates 
-The templates in this folder are made for SugarCube 2.36.1. Any use with a previous or later version may break the code.\
-The HTML template file can be opened in any Twine 2 version from 2.36.1 onward.
-> All the templates will be updated for 2.37, a copy of the 2.36.1 will be kept available.
+The templates in this folder are made for SugarCube 2.36.1. The HTML template file can be opened in any version Twine 2 that includes 2.36.1.\
+Using those templates with any older or newer version of SugarCube may break the code.
+> All the templates will be updated for 2.37 (TBD), a copy of the 2.36.1 will still be available here.
 
 The templates are organised in 2 categories:
 * Code Templates (settings, character creator)
 * UI Templates (interface change)
+
+You can find a more comprehensive list of my Harlowe templates, along with their specs [here](SugarCube/DocsSugarCube.md).
 
 ## Ready-to-Use Tweego Folder
 This *Ready-to-Use Tweego* folder is not of the easiest (if not **the** easiest) install for Tweego.\
@@ -36,4 +53,16 @@ Just:
 
 -> You're ready to code!
 
-The Folder also includes a Guide for use - based on a use with the VSCode editor.
+The Folder also includes a Guide for use - based on a use with the VSCode editor. A "playable" version of the Guide is available [here](https://manonamora.itch.io/ready-to-use-tweego-folder).
+
+You can find a more comprehensive list of specs of this Folder [here](Tweego%20Folder/DocsTweego.md).
+
+## Links
+You can also find these templates on my [itch.io page](https://manonamora.itch.io/), specifically:
+
+* [100% Good Twine Harlowe Templates](https://manonamora.itch.io/twine-harlowe-templates)
+* [100% Good Twine SugarCube Templates](https://manonamora.itch.io/twine-sugarcube-templates)
+* [100% Good Ready-to-Use Tweego Folder](https://manonamora.itch.io/ready-to-use-tweego-folder)
+* [Twine Templates, Guides and Code](https://itch.io/c/2917452/twine-templates-guides-and-code)
+
+> For more Twine resources, see the [Twine Resource Masterlist](https://intfiction.org/t/twine-resource-masterlist-wiki/65903) on the IntFiction Forum.
